+++ shm/Game.hpp
@@ -29,7 +29,7 @@ private:
 
     void travel();
     Coordinates getTravelLocation();
-    void advanceTimeTraveling(size_t distance);
+    void advanceTimeTraveling(int distance);
     void printOptions() const;
     void makeAction(Action choice);
     void exit() const;
